DateTime und ratio als flow mit korrekten Dezimalstellen implementiert.

+++ app/src/main/java/marcelfuchs/example/org/keed/NewItemFragment.kt
@@ -6,7 +6,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import kotlinx.android.synthetic.main.fragment_list.*
 import kotlinx.android.synthetic.main.fragment_new_item.*
 
 class NewItemFragment : Fragment() {
@@ -31,8 +30,8 @@ class NewItemFragment : Fragment() {
 
     fun addItem() {
         // get Strings from the editText TextViews
-        val kill = et_enter_kills.text.toString().toInt()
-        val death = et_enter_deaths.text.toString().toInt()
+        val kill = et_enter_kills.text.toString().toFloat()
+        val death = et_enter_deaths.text.toString().toFloat()
 
         val newItem = Keed(kill, death)
         mKeedList.add(newItem)
